Finished moving delegation queries into their own module
